Add files via upload
The keyboard images are provided. The program can measure the size of the keyboard's length and its breadth. I try to match the feature from the missing key to the frame so that it can detect the missing keys and also identify the type of keyboard. we are provided with 2 keyboards so the program can identify between 2 keyboards.